[FLINK-19073][table-planner] Imporve streamExecTemporalJoinRule for stream
