+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/plan/rules/physical/stream/StreamExecLegacyTemporalJoinRule.scala
@@ -18,48 +18,57 @@
 
 package org.apache.flink.table.planner.plan.rules.physical.stream
 
-import java.util
-
 import org.apache.flink.table.planner.plan.`trait`.FlinkRelDistribution
 import org.apache.flink.table.planner.plan.nodes.FlinkConventions
 import org.apache.flink.table.planner.plan.nodes.logical._
-import org.apache.flink.table.planner.plan.nodes.physical.stream.StreamExecTemporalJoin
-import org.apache.flink.table.planner.plan.utils.TemporalJoinUtil
-
+import org.apache.flink.table.planner.plan.nodes.physical.stream.StreamExecLegacyTemporalJoin
+import org.apache.flink.table.planner.plan.utils.{FlinkRelOptUtil, IntervalJoinUtil, LegacyTemporalJoinUtil}
 import org.apache.calcite.plan.RelOptRule.{any, operand}
 import org.apache.calcite.plan.{RelOptRule, RelOptRuleCall, RelTraitSet}
 import org.apache.calcite.rel.RelNode
 import org.apache.calcite.rel.core.JoinRelType
+import java.util
 
 /**
- * Rule that matches a temporal join node and converts it to [[StreamExecTemporalJoin]],
- * the temporal join node is a [[FlinkLogicalJoin]] whose left input is a [[FlinkLogicalRel]]
- * and right input is a [[FlinkLogicalSnapshot]].
+ * Rule that matches a legacy temporal join node which rewritten by
+ * [[LEGACY_TEMPORAL_JOIN_CONDITION]], and converts it to [[StreamExecLegacyTemporalJoin]].
  */
-class StreamExecTemporalJoinRule
+class StreamExecLegacyTemporalJoinRule
   extends RelOptRule(
-    operand(classOf[FlinkLogicalJoin],
+    operand(
+      classOf[FlinkLogicalJoin],
       operand(classOf[FlinkLogicalRel], any()),
-      operand(classOf[FlinkLogicalSnapshot],
-        operand(classOf[FlinkLogicalRel], any()))),
-    "StreamExecTemporalJoinRule") {
+      operand(classOf[FlinkLogicalRel], any())),
+    "StreamExecLegacyTemporalJoinRule") {
 
   override def matches(call: RelOptRuleCall): Boolean = {
     val join = call.rel[FlinkLogicalJoin](0)
-    val isTemporalJoin = TemporalJoinUtil.containsTemporalJoinCondition(join.getCondition)
+    val joinInfo = join.analyzeCondition
 
-    val supportedJoinTypes = Seq(JoinRelType.INNER, JoinRelType.LEFT)
-    isTemporalJoin && supportedJoinTypes.contains(join.getJoinType)
+    if (!LegacyTemporalJoinUtil.containsTemporalJoinCondition(join.getCondition)) {
+      return false
+    }
+
+    val tableConfig = FlinkRelOptUtil.getTableConfigFromContext(join)
+    val (windowBounds, _) = IntervalJoinUtil.extractWindowBoundsFromPredicate(
+      joinInfo.getRemaining(join.getCluster.getRexBuilder),
+      join.getLeft.getRowType.getFieldCount,
+      join.getRowType,
+      join.getCluster.getRexBuilder,
+      tableConfig)
+
+    windowBounds.isEmpty && join.getJoinType == JoinRelType.INNER
   }
 
   override def onMatch(call: RelOptRuleCall): Unit = {
     val join = call.rel[FlinkLogicalJoin](0)
     val left = call.rel[FlinkLogicalRel](1)
-    val snapshotInput = call.rel[FlinkLogicalRel](3)
+    val right = call.rel[FlinkLogicalRel](2)
 
-    def toHashTraitByColumns(
-        columns: util.Collection[_ <: Number],
-        inputTraitSets: RelTraitSet) = {
+    val traitSet: RelTraitSet = join.getTraitSet.replace(FlinkConventions.STREAM_PHYSICAL)
+    val joinInfo = join.analyzeCondition
+
+    def toHashTraitByColumns(columns: util.Collection[_ <: Number], inputTraitSets: RelTraitSet) = {
       val distribution = if (columns.size() == 0) {
         FlinkRelDistribution.SINGLETON
       } else {
@@ -69,19 +78,17 @@ class StreamExecTemporalJoinRule
         replace(FlinkConventions.STREAM_PHYSICAL).
         replace(distribution)
     }
-
-    val joinInfo = join.analyzeCondition
     val (leftRequiredTrait, rightRequiredTrait) = (
       toHashTraitByColumns(joinInfo.leftKeys, left.getTraitSet),
-      toHashTraitByColumns(joinInfo.rightKeys, snapshotInput.getTraitSet))
+      toHashTraitByColumns(joinInfo.rightKeys, right.getTraitSet))
 
     val convLeft: RelNode = RelOptRule.convert(left, leftRequiredTrait)
-    val convRight: RelNode = RelOptRule.convert(snapshotInput, rightRequiredTrait)
-    val providedTraitSet: RelTraitSet = join.getTraitSet.replace(FlinkConventions.STREAM_PHYSICAL)
+    val convRight: RelNode = RelOptRule.convert(right, rightRequiredTrait)
 
-    val temporalJoin = new StreamExecTemporalJoin(
+
+    val temporalJoin = new StreamExecLegacyTemporalJoin(
       join.getCluster,
-      providedTraitSet,
+      traitSet,
       convLeft,
       convRight,
       join.getCondition,
@@ -91,6 +98,6 @@ class StreamExecTemporalJoinRule
   }
 }
 
-object StreamExecTemporalJoinRule {
-  val INSTANCE: RelOptRule = new StreamExecTemporalJoinRule
+object StreamExecLegacyTemporalJoinRule {
+  val INSTANCE: RelOptRule = new StreamExecLegacyTemporalJoinRule
 }
