+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/plan/utils/LegacyTemporalJoinUtil.scala
@@ -1,0 +1,106 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.flink.table.planner.plan.utils
+
+import org.apache.flink.util.Preconditions.checkArgument
+
+import org.apache.calcite.rex._
+import org.apache.calcite.sql.`type`.{OperandTypes, ReturnTypes}
+import org.apache.calcite.sql.{SqlFunction, SqlFunctionCategory, SqlKind}
+
+/**
+  * Utilities for legacy temporal table join from temporal table function.
+  */
+object LegacyTemporalJoinUtil {
+
+  // ----------------------------------------------------------------------------------------
+  //                          Temporal TableFunction Join Utilities
+  // ----------------------------------------------------------------------------------------
+
+  /**
+    * [[LEGACY_TEMPORAL_JOIN_CONDITION]] is a specific condition which correctly defines
+    * references to rightTimeAttribute, rightPrimaryKeyExpression and leftTimeAttribute.
+    * The condition is used to mark this is a temporal table function join.
+    * Later rightTimeAttribute, rightPrimaryKeyExpression and leftTimeAttribute will be
+    * extracted from the condition.
+    */
+  val LEGACY_TEMPORAL_JOIN_CONDITION = new SqlFunction(
+    "__LEGACY_TEMPORAL_JOIN_CONDITION",
+    SqlKind.OTHER_FUNCTION,
+    ReturnTypes.BOOLEAN_NOT_NULL,
+    null,
+    OperandTypes.or(
+      OperandTypes.sequence(
+        "'(LEFT_TIME_ATTRIBUTE, RIGHT_TIME_ATTRIBUTE, PRIMARY_KEY)'",
+        OperandTypes.DATETIME,
+        OperandTypes.DATETIME,
+        OperandTypes.ANY),
+      OperandTypes.sequence(
+        "'(LEFT_TIME_ATTRIBUTE, PRIMARY_KEY)'",
+        OperandTypes.DATETIME,
+        OperandTypes.ANY)),
+    SqlFunctionCategory.SYSTEM)
+
+  def isRowtimeCall(call: RexCall): Boolean = {
+    checkArgument(call.getOperator == LEGACY_TEMPORAL_JOIN_CONDITION)
+    call.getOperands.size() == 3
+  }
+
+  def isProctimeCall(call: RexCall): Boolean = {
+    checkArgument(call.getOperator == LEGACY_TEMPORAL_JOIN_CONDITION)
+    call.getOperands.size() == 2
+  }
+
+  def makeRowTimeTemporalJoinConditionCall(
+    rexBuilder: RexBuilder,
+    leftTimeAttribute: RexNode,
+    rightTimeAttribute: RexNode,
+    rightPrimaryKeyExpression: RexNode): RexNode = {
+    rexBuilder.makeCall(
+      LEGACY_TEMPORAL_JOIN_CONDITION,
+      leftTimeAttribute,
+      rightTimeAttribute,
+      rightPrimaryKeyExpression)
+  }
+
+  def makeProcTimeTemporalJoinConditionCall(
+    rexBuilder: RexBuilder,
+    leftTimeAttribute: RexNode,
+    rightPrimaryKeyExpression: RexNode): RexNode = {
+    rexBuilder.makeCall(
+      LEGACY_TEMPORAL_JOIN_CONDITION,
+      leftTimeAttribute,
+      rightPrimaryKeyExpression)
+  }
+
+  def containsTemporalJoinCondition(condition: RexNode): Boolean = {
+    var hasTemporalJoinCondition: Boolean = false
+    condition.accept(new RexVisitorImpl[Void](true) {
+      override def visitCall(call: RexCall): Void = {
+        if (call.getOperator != LEGACY_TEMPORAL_JOIN_CONDITION) {
+          super.visitCall(call)
+        } else {
+          hasTemporalJoinCondition = true
+          null
+        }
+      }
+    })
+    hasTemporalJoinCondition
+  }
+
+}
